+++ mdparser_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit mdparser.stub.php instead.
- * Stub hash: 675af44e3bb6a21d01b7024f08469d344ffc1233 */
+ * Stub hash: 4b8fb16f2c86ca2a69272cff11d938a51f274c52 */
 
 ZEND_BEGIN_ARG_INFO_EX(arginfo_class_MdParser_Options___construct, 0, 0, 0)
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, sourcepos, _IS_BOOL, 0, "false")
@@ -31,6 +31,7 @@ ZEND_BEGIN_ARG_INFO_EX(arginfo_class_MdParser_Options___construct, 0, 0, 0)
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, spoilers, _IS_BOOL, 0, "false")
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, latexMath, _IS_BOOL, 0, "false")
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, wikiLinks, _IS_BOOL, 0, "false")
+	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, admonitions, _IS_BOOL, 0, "false")
 ZEND_END_ARG_INFO()
 
 ZEND_BEGIN_ARG_WITH_RETURN_OBJ_INFO_EX(arginfo_class_MdParser_Options_strict, 0, 0, MdParser\\Options, 0)
@@ -285,6 +286,12 @@ static zend_class_entry *register_class_MdParser_Options(void)
 	zend_string *property_wikiLinks_name = zend_string_init("wikiLinks", sizeof("wikiLinks") - 1, true);
 	zend_declare_typed_property(class_entry, property_wikiLinks_name, &property_wikiLinks_default_value, ZEND_ACC_PUBLIC|ZEND_ACC_READONLY, NULL, (zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_BOOL));
 	zend_string_release_ex(property_wikiLinks_name, true);
+
+	zval property_admonitions_default_value;
+	ZVAL_UNDEF(&property_admonitions_default_value);
+	zend_string *property_admonitions_name = zend_string_init("admonitions", sizeof("admonitions") - 1, true);
+	zend_declare_typed_property(class_entry, property_admonitions_name, &property_admonitions_default_value, ZEND_ACC_PUBLIC|ZEND_ACC_READONLY, NULL, (zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_BOOL));
+	zend_string_release_ex(property_admonitions_name, true);
 
 	return class_entry;
 }