+++ mdparser_options.c
@@ -71,6 +71,7 @@ enum {
     MDOPT_SPOILERS,
     MDOPT_LATEX_MATH,
     MDOPT_WIKILINKS,
+    MDOPT_ADMONITIONS,
     MDOPT_COUNT_
 };
 
@@ -120,6 +121,7 @@ static const mdparser_options_field mdparser_options_fields[] = {
     F("spoilers",                   MD_FLAG_SPOILERS, 0, false),
     F("latexMath",                  MD_FLAG_LATEXMATHSPANS, 0, false),
     F("wikiLinks",                  MD_FLAG_WIKILINKS, 0, false),
+    F("admonitions",                MD_FLAG_ADMONITIONS, 0, false),
 };
 
 #undef F
@@ -279,6 +281,7 @@ PHP_METHOD(MdParser_Options, __construct)
         Z_PARAM_BOOL(values[MDOPT_SPOILERS])
         Z_PARAM_BOOL(values[MDOPT_LATEX_MATH])
         Z_PARAM_BOOL(values[MDOPT_WIKILINKS])
+        Z_PARAM_BOOL(values[MDOPT_ADMONITIONS])
     ZEND_PARSE_PARAMETERS_END();
 
     mdparser_options_populate_object(Z_OBJ_P(ZEND_THIS), values);